don't cast, think. then look. then, don't cast ;-)
svn path=/trunk/extragear/sysadmin/partitionmanager/; revision=1107142

+++ src/core/copysourceshred.cpp
@@ -31,7 +31,7 @@ CopySourceShred::CopySourceShred (qint64 s, qint32 sectorsize) :
 	CopySource(),
 	m_Size(s),
 	m_SectorSize(sectorsize),
-	m_SourceFile(Config::shredSource == static_cast<int>(Config::EnumShredSource::random) ? "/dev/urandom" : "/dev/zero")
+	m_SourceFile(Config::shredSource() == Config::EnumShredSource::random ? "/dev/urandom" : "/dev/zero")
 {
 }
